+++ src/Core/EBF_I2C.h
@@ -20,7 +20,7 @@ class EBF_I2C : protected EBF_HalInstance, public Stream {
 
 		uint8_t Process();
 
-		void SetPollInterval(uint16_t ms);
+		void SetPollInterval(uint32_t ms);
 
 		// Stream and Print interfaces
 		inline size_t write(unsigned long n) { return i2c.write((uint8_t)n); }
