+++ src/Products/EBF_HC-SR04_DistanceSensor.h
@@ -17,7 +17,7 @@ class EBF_HCSR04_DistanceSensor : protected EBF_DigitalInput {
 		float GetLastValue() { return lastValue; }
 
 		// Setting polling interval in milli-seconds
-		void SetPollInterval(uint16_t ms) { pollIntervalMs = ms; }
+		void SetPollInterval(uint32_t ms) { pollIntervalMs = ms; }
 
 		void SetOnChange(EBF_CallbackType onChangeCallback, uint8_t changePercent = 5);
